students/K33391/Akulov_Aleksey/Lab_3/parser/requirements.txt
txt
requests
beautifulsoup4
psycopg2-binary
fastapi
pydantic
uvicorn
sqlalchemy